+++ .github/workflows/update-tokens.yml
@@ -29,10 +29,10 @@ jobs:
         run: yarn postinstall
       - name: Test implementation
         run: yarn test
-      - name: add Figma token to .env
-        run: echo "FIGMA_TOKEN=${{ secrets.FIGMA_IKONER_API }}" >> @navikt/core/tokens/.env
       - name: Update tokens
         run: yarn update:tokens
+        env:
+          FIGMA_TOKEN: ${{ secrets.FIGMA_IKONER_API }}
       - name: Create Pull Request
         uses: peter-evans/create-pull-request@v3
         with:
