+++ @navikt/core/tokens/figma/fetch.ts
@@ -1,7 +1,9 @@
 import Axios from "axios";
 const { create } = Axios;
+const dotenv = require("dotenv");
+dotenv.config();
 
-const FIGMA_FILE = "A7fbghTh75Hk65SieEhaKK";
+const FIGMA_FILE = "m0qpSPpsnNC9DwdmerGqJH";
 
 const FigmaAxion = (token) =>
   create({
@@ -10,11 +12,24 @@ const FigmaAxion = (token) =>
     },
   });
 
-export const getSyncDocument = async () => {
+export const fetchFile = async () => {
   const { data } = await FigmaAxion(process.env.FIGMA_TOKEN)
     .get(`https://api.figma.com/v1/files/${FIGMA_FILE}`)
     .catch((e) => {
       throw e;
     });
   return data;
+};
+
+export const fetchFileStyles = async () => {
+  const {
+    data: {
+      meta: { styles },
+    },
+  } = await FigmaAxion(process.env.FIGMA_TOKEN)
+    .get(`https://api.figma.com/v1/files/${FIGMA_FILE}/styles`)
+    .catch((e) => {
+      throw e;
+    });
+  return styles;
 };
