+++ @navikt/core/tokens/package.json
@@ -21,8 +21,7 @@
     "watch": "nodemon --watch src/index.js --exec \"yarn build\"",
     "token": "style-dictionary build",
     "test": "jest",
-    "update": "ts-node figma/index.ts dotenv/config",
-    "figma": "ts-node figma/temp.ts dotenv/config"
+    "update": "ts-node figma/index.ts dotenv/config"
   },
   "repository": {
     "type": "git",
